fix precision on csv output
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@36088 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objects/omssa/MSResponse.cpp
@@ -72,7 +72,7 @@ void CMSResponse::PrintCSV(ostream& os, CRef <CMSModSpecSet> ModSet)
     
     CMSResponse::THitsets::const_iterator iHits;
     iHits = GetHitsets().begin();
-	float fMZ_scale = CanGetScale() ? GetScale() :  100.0;
+	double fMZ_scale = CanGetScale() ? GetScale() :  100.0;
     for(; iHits != GetHitsets().end(); iHits++)  
 	{
 		CRef< CMSHitSet > HitSet =  *iHits;
